cart component created

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -2,7 +2,7 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
-
+const { hasPermission } = require('../utils');
 const { transport, makeANiceEmail } = require('../mail');
 
 const Mutations = {
@@ -43,9 +43,16 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id title }`)
+    const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
     // 2. check if they own that item, or have the permissions
-    // TO DO
+    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermission = ctx.request.user.permissions.some
+      (permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+
+    if (!ownsItem || !hasPermission) {
+      throw new Error('You do not have permission to do that')
+    }
     // 3. Delete it
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -59,7 +66,7 @@ const Mutations = {
       data: {
         ...args,
         password,
-        permissions: { set: ['USER'] }
+        permissions: { set: ['ADMIN'] }
       }
     }, info);
     // create the JWT token for them
@@ -170,6 +177,33 @@ const Mutations = {
     });
     // return the new user
     return updatedUser;
+  },
+
+  async updatePermissions(parent, args, ctx, info) {
+    // 1. Check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error('you must be logged in')
+    }
+    // 2. Query the current user
+    const currentUser = await ctx.db.query.user({
+      where: {
+        id: ctx.request.userId,
+
+      }
+    }, info);
+    // 3. Check if they have permissions to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE'])
+    // 4. update the permission
+    return ctx.db.mutation.updateUser({
+      data: {
+        permissions: {
+          set: args.permissions,
+        }
+      },
+      where: {
+        id: args.userId
+      }
+    }, info);
   },
 
 };
